Avoid need to inject BeanDefinitionRepairer everywhere as a helper (#5365)
* Avoid need to inject BeanDefinitionRepairer everywhere as a helper.

Instead we record the bean classes of the initial definitions as they are registered,
so we can repair them as necessary later on. This also means we can remove the extra
call to 'setBeanClassName' which was overwriting the resolved class in the definition
(they share a single field) and was actually making the original issue more likely.

This approach works better in OSGi where otherwise we might end up injecting multiple
BeanDefinitionRepairer helpers, which increases the risk of a type being registered in
a repairer class that's different to the one actually used to repair definitions.