Ensure safe class loading in instrumentation
Loading a class during transformation/instrumentation MUST not be the
current class being transformed otherwise a LinkageError happen with
message "attempted duplicate class definition" because we have loaded
the class before it was transformed, so when transformation finish
the original load of the class is happening and boom!
We introduced a helper method that require to provide the name of
the class currently being transformed to ensure it is not the same.
We also catch Throwable in transform method to ensure whatever happens
during transformation we do not bubble up to the user code and load
the original class