Add special support for enum values
Enum in java are in fact regular classes with a special treatment by
the java compiler. from the POV of the JVM there is no difference.
From user perspective however it would be nice to have constant name
as enum values.
We are serializing then the enum constant name in the snapshot instead
of handling enum as regular complex class