Fix log template issue for duplicated line probes
Having 2 line probes on the same location with same template or
accessing the same field, arg or var result in an error evaluating
the template.
With line probe, the freeze/commit of the data is done just after the
evaluation per probe id.
The fix is to perform 2 loops on the probe ids, first only evaluation
and then commit.