[Spark] Add support for spark structured streaming (#5629)
Spark structured streaming is a stream processing engine built on the Spark SQL. It process data streams as a series of small batch jobs, where each batch jobs is composed of the regular spark operations (job / stage / task)

A streaming applications is composed of:
- one or more long running queries, identified a queryId
- each query continuously executes batches identified by a incrementing batchId

Spark is emitting 3 type of events in structured streaming:
- QueryStartedEvent
- QueryTerminatedEvent
- QueryProgressEvent

QueryProgressEvent is emitted at the end of each micro-batch in the finishTrigger method and has a lot of metrics about the executed batch

There is no event emitted at the start of a micro batch (QueryStartedEvent is emitted at the start of the long running query, not for every micro-batch). To circumvent that, the batch span is started at the start of its first job, the caveat is that the first job is started after the getBatch and queryPlanning computations, leading to the batch span being shorter that the real batch duration

Spark structured streaming is part of the spark-sql module, meaning that the muzzle directive has to be changed. This is acceptable as Spark SQL is the recommended way to use spark and vendors like EMR or databricks always provide spark-sql in their managed versions