Enable ByteCode verification by default
Once we have generated the new classfile we run a static analysis of
the bytecode using the ASM library's analyzer. It helps find out
before sending back the modified classfile to the JVM if the bytecode
is not correct and then cancelling the instrumentation to not impact
the user code.