Skip collection of config during native-image build (since telemetry is off)
We already do this for `InstrumenterConfig.SINGLETON` which is used during instrumentation.

While we avoid touching `Config` during the build, we mark it as "rerun" in  our patched `native-image` config.
The `native-image` generator will analyze it indirectly to check what's needed to "rerun" class initialization,
so we should therefore have the same check to turn off collection in `Config.SINGLETON`.