Use alpine base for lib-injection image
`libdl.so.2: cannot open shared object file: No such file or directory`
errors have been reported in customer environments which prevents
application containers from starting up when using [library
injection](https://docs.datadoghq.com/tracing/trace_collection/library_injection_local/?tab=kubernetes#step-3---tag-your-pods-with-unified-service-tags).
Likely this is due to some other software such as instrumentation or
security that modifies the command run in the container.

`busybox` [lacks many standard C
libraries](https://github.com/docker-library/busybox/issues/46).

`alpine` is a similarly sized image which does provide the standard C
libraries which mitigates the issue

There is little risk in changing the base image to `alpine` as the only
command required is to copy files from the image to a volume. It is
possible that the command that is being added to the container is not
compatible with alpine images.

Change the base image from `busybox` to `alpine` to make the image more
compatible with other software environments.