Rename operation name spark.batch to spark.streaming_batch (#5953)
Rename spark.batch operation name to spark.streaming_batch. This is a breaking change that is already done in the backend.

It’s not clear that spark.batch is a spark stream operation. By using streaming_batch we match the spark streaming listener terminology and make it clear it’s a stream