Truncate long String values in IAST Vulnerabilities (#5975)
What Does This Do
When a source or a value part has value or pattern that exceed configured max size, truncate it to the max size and add a "truncated" attribute with the truncation direction as value (for this iteration it's always "right")

Motivation
Avoid vulnerabilities report with very long texts. For example, when the customer application is saving blog posts in a database, this would generate a very big evidence with thousands of chars in a single source and valuePart.

This kind of vulnerabilities would impact the used network bandwidth, serialization time in the library, unserialization time in backend and the visualization of a vulnerability with thousands of chars.