Make sure internal 'datadog.trace.api.Functions' class is relocated in dd-trace-ot
DDTracer uses Functions.UTF8_ENCODE when creating spans. 'Functions' is currently
excluded from relocation while 'UTF8BytesString' which it depends on is relocated.
This can lead to class-cast issues when attaching 'dd-java-agent' and 'dd-trace-ot'
even when automatic tracing is explicitly disabled.

A similar issue can occur with the Endpoint internal API so we also relocate that.