Fix IAST ranges for StringBuffer.toString
A StringBuffer or StringBuilder object might be tainted directly. In
that case, we'll use an open Range (signaling that the full object is
tainted, without a specific length). On StringBuffer#toString, this will
lead to a String with an open Range, which the rest of our code base
does not expect.