Improve snapshot pruning algorithm
Instead of decreasing the depth of the collect by re-parsing the
snapshot with the Json parser, we are scanning the the json string and
building a tree structure representing the json structure with indexed
string segments. WE have special treatment for notCapturedReason nodes
that helps us in the pruning algorithm:
Leaves of this structure are retrieved and then sorted through a
priority queue where the sort key is the following:
 - notCapturedReason depth node
 - level
 - notCapturedReason
 - size of the node
Then following this, we are pruning first leaves with
notCapturedReason depth then removing deepest level leaves,
notCapturedReason leaves and finally the leaves with largest size.
if the leaves of the original snapshot are not enough for reducing
the size we are considering parents.
Nodes pruned are replaced by `{"pruned":true}`.