Send probe statuses to new debugger track
datadog agent is now exposing a new endpoint: /debugger/v1/diagnostics
This endpoint is used to send probe statuses to a dedicated debugger
track.
we add detection o this new endpoint and create a dedicated
BatchUploader to this endpoint using attachment.
if the endpoint is not available (old datadog agent) we send to the
previous url as before.