Add emitting status for probes
Add a new status for probes to indicate the fact that they are sending
data. a new EMITTING status is added and for each probe implementation
we send this status.
- Log probes: when snapshot is sent
- metric probes: when we sent the metric to statsd
- span probes: when the span finishes
- span decoration probes: when all tags were added
The status is set and sent only the first time.
Refactoring was done for providing full probe id (uuid + version)
inside the probe evaluation. For this, it required to inject uuid
(ProbeId) and resolved by using ConfigurationUpdater.
Refactored also smoke tests to test the status is reached when data
are sent.