Fix race in Spring Security instrumentation (#6277)
Stop relying on changing hideUserNotFoundExceptions, which can break
applications by changing their observed behavior.

We will detect user not found when UsernameNotFoundException is created
(which then might be re-thrown or changed into BadCredentialsException).

This results in a partial loss of functionality: we will not set the tag
for user exists to true. Only to false. When we cannot reliably
determine if the user exists or not, we will simply not set
`appsec.events.users.login.failure.usr.exists`. This might be revisited
in the future.

Solves #6072.