Enforce ordering of probe instrumentation
To make sure on a same method all probe defined on this location are
instrumented in right order to interact each other. Order of
instrumentation is:
 - metric probes
 - log probes
 - span decoration probes
 - span probes
This way the last probe wraps method body and all other
instrumented code.
Span should be first to be executed to create a span for the current
method then span decoration should add tags for this new span. Logs
should be decorated with trace/span ids for the current method.
Finally metric are last because no dependency.