Support special case of empty catch block
if you want to put a line probe in an empty catch block, exception
variable is not declared in that special case, so to capture it
as local var we need to explicitly add it. We declare the missing
local var if not present and store the exception present on the stack
in it. then we collect as usual the local vars and the exception
will be then present in the snapshot