Fix JSONObject instrumentation advice
Instrumentation for `JSONObject#get` in IAST is not using an inlined
advice. This seems to be causing a NoClassDefFoundError. This only
happens when `DD_IAST_ENABLED=true` and has no impact in the application
since it is caught by our mechanism for unhandled exceptions in
instrumentation. It might have had some performance impact in
services with `DD_IAST_ENABLED=true` and intensive use of
`JSONObject#get`.