Add instrumentation of exception stacktraces
When instrumenting exception on first occurence, we get the stacktrace
of the exception and for each frame we will generate a special method
probe (ExceptionProbe) based on LogProbe doing a snapshot at Exit.
to place correctly the probes we are basing on StackTrace element that
gives class, method and line number.
for overloaded methods (with same name) the only way to disambiguate
which method to instrument is to use the line number. We introduce
then the ClassFileLines class to generalize the LineNumber information
extracted from the parsing of a classfile by the ASM library and be
reused for line instrumentation and method matching. previous LineMap
class is then replaced by it.
Created exception probe are kept by ExceptionProbeManager and must be
merged with the Configuration coming from Remote Config. Therefore
the ConfigurationUpdater class must now take into account that and
also be able to be triggered not only by the Remote Config thread
but also the thread coming from the exception captured added to the
span. Applying configuration require now to be atomic between
Configuration object, new transformer and retransformation of the
classes impacted.