Fixes to trace agent service rates feedback - e.g. MAX TPS support
This change set fixes the MAX TPS support in the tracer.
Previously, the tracer treated service name and env in a case sensitive fashion.
Additionally, the tracer didn't properly handle the fallback case of an empty service and empty env.

While the empty service and empty env case was parsed properly, getSampler would return a static DEFAULT sampler that always used a rate 1.0.  This wasn't obvious from the tests because the tests also use a rate of 1.0.