Add frame tags and send snapshots on addThrowable
Collect snapshot generated by an exception probe frame in
ThrowableState class that is generated by exception instance. Those
exception instances are tracked by a WeakIdentityHashMap in
ExceptionProbeManager. Key is directly the exception instance. The
Weak nature of the hash map allow to remove the entry upon the
exception is no longer reachable.

When unwinding, once the exception reaches the frame level where it
is added to the span (DDSpan::addThrowable) we process all the
collected snapshots and and add tags with frame index for each
snapshot id. The frame index is computed based on the generated
mapping between original exception (the one added to the span) and
the inner most exception which was used to generated exception probes
and snapshots.

An instrumentation bug is also fixed:
casting when calling CapturedValue.of is not necessary as the method
is accepting any objects. On the other side, some return types could
not be used as cast target because not visible from the current frame.
inherited field with a type not visible from the inherited class.