Decouple ConfigurationUpdater and exception probes
Introducing the concept of source for ConfigurationUpdater that is
just a list of ProbeDefinition. As RemoteConfig configuration are only
 for probes for now, we are reducing it to this list. Hence the accept
method in ConfigurationUpdater is taking a source
(REMOTE_CONFIG or EXCEPTION) and a list of ProbeDefinition.
For RemoteConfig part, we are building this list instead of
Configuration object and capping the umber of probes per type at that
level now instead of ConfigurationUpdater. The side effect of this
is that now there is no capping for probe definitions coming from file.
Configuration::accept method is now responsible to merge definitions
from all source (REMOTE_CONFIG and EXCEPTION) and create a
Configuration object that will be applied as before.
As a consequence, ExceptionProbeManager can now be created directly
into DefaultExceptionDebugger class.