Refactor ExceptionProbe to get one per frame
Instead of having multiple probes for each frame when sharing same
frame for different exception (and fingerprint) we have only one
instance of ExceptionProbe per frame and using ExceptionProbeManager
we make sure that:
1. we only instrument once.
2. we compute the fingerprint and assess if we need to capture the
exception or not for this frame
When instrumenting, we store a normalized version of a Where instance
with className + methodName + signature. The normalization can only
happen at instrumentation time to resolve which line correspond to
which method (in case of overloaded methods).
When retransformations happen for a class we need to remove the
instrumented method from ExceptionProbeManager by introducing a
listener mechanism in ConfigurationUpdater.