Fix inaccessible fields capture
NoSuchFieldException can be thrown if we try to capture, through
reflection, (static) fields from JDK classes or from another module.
To fix this we need to catch eventual exception thrown by accessing
field values.
For that we continue to try to access directly to field when possible
because it will be faster, but when we detect fields are not
accessible we use another method that try to access it through
reflection but could catch exception if not and returns a
CapturedValue instance filled with the reason of this exception.