Add Throwable capturing fields support for JDK16+
Since JDK16 it's not possible to extract values by reflection for JDK
modules without adding `--add-opens=java.base/java.lang=ALL-UNNAMED`
so when capturing an exception (like in @exception) we are not allowed
to capture fields like detailMessage, cause, stacktrace which can be
valuable to troubleshoot.
We introduce in WellKnownClasses the mechanism to fetch fields through
dedicated call to getter for registered classes.
In the special case of Throwable we are allowing any inherited classes
to capture the Throwable fields as well as custom fields. Though to
capture Throwable fields (like detailMessage) we need to call getter
that could be overridden. Hence we are checking before calling them
that they are not overridden by checking the declaring class of the
getter.
Add option in build to not allow reflection access for debugger tests
to assess that we can access correctly even for recent JDKs