Move the OpenTelemetry shim to its own module so we can re-use it for drop-in support.
We now embed a shaded copy of the OpenTelemetry API which has been pre-instrumented
at build time with our shim. This shaded API will be on the bootstrap class-path so it
can be used by drop-in advice without having to continually inject it everywhere the
drop-in advice is injected.

The same shim is used at runtime when instrumenting use of the OpenTelemetry API on
the application class-path.