Add high rate queue for log template snapshots
Historically, the Queue inside SnapshotSink was designed for a low
rate number of snapshots (1 per second) and the consumer thread is
the one shared with all other tasks (AgentTaskScheduler.INSTANCE). So
the reactivity of the thread to consume the queue is not enough for
handling up to 5000 snapshots/s allowed for log template probes.
We introduce then a dedicated high rate queue with dedicated threads
with better reactivity (max 100ms between polls). All non-capturing
snapshots are added into this queue.