Remove explicit capture of fields
Previously we were capturing directly fields of the current class in
a specific way and use them for condition evaluation, while this
was captured as argument. Now we only capture this and all evaluations
are based on implicit this dereferencing when needed.
NB: Serialization of complex classes no longer capture static fields