Adding "InitializationTelemetry"
Adding InitializationTelemetry to the bootstrap

InitializationTelemetry is used to report guard rail check failures via a helper process installed by single step

Most of the real changes are in AgentBootstrap with a lot of attention towards working with restrictive SecurityManagers.  With the most restrictive SecurityManagers access to env & properties are blocked, so SecurityPermission & Throwable checks are usually liberally throughout.

For constructing the JSON that goes to the helper process, we don't want to introduce dependencies into the bootstrap.  So I've implemented a small JsonBuffer class modeled loosely after GSON's JsonWriter.