Adding SecurityManager tests
Adding tests to make sure that AgentBootstrap doesn't fail even with a very strict SecurityManager

These tests are implemented via SecurityManagerCheck which is a simple main designed to be used with a javaagent and a TestSecurityManager

TestSecurityManager is a SecurityManager with a minimal set of permissions for the javaagent to run normally

TestSecurityManager has various sub-classes that restrict access to different things used by the javaagent.  The aim is to make that under no circumstance does the javaagent fail in a way that prevents the program from running normally.

More tests to follow