Add Parameter to only inject data jobs for particular java commands (#7366)
Add parameter `dd.data.jobs.command.pattern` to only inject the tracer on specific command pattern for Data Jobs

In the context of Data Jobs Monitoring, change the Java tracer to become more selective about which Java process to instrument when Data Jobs is enabled and a Java command regex pattern is specified. In runtime environment where we are not sure the command pattern to use, the java tracer should just behave as before by continuing the Agent start.

With single step instrumentation, Java tracer is injected for all Java processes, this is causing some initial system overhead as Java tracer starts in those JVMs. In resource limited environment, this can cause the host to be replaced due to high system load. For Data Jobs Monitoring, the Java tracer only cares about Spark process running the driver and executors, this newly added command pattern allows us to exit the Agent early if we don't think the process is Spark related and saving the overhead of increased System load.

- The environment variable to specify the data jobs Java command pattern (regex):DD_DATA_JOBS_COMMAND_PATTERN.
- DD_DATA_JOBS_COMMAND_PATTERN only takes effects if DD_DATA_JOBS_ENABLED is set to true.
---------

Co-authored-by: Liangzhou Yi <1934063+yiliangzhou@users.noreply.github.com>