Fix SymDB upload dropped requests
SymbolSink was emitting an upload requests for every scope added into
the sink queue. As the queue was 1024 entries in capacity we could
easily end up in a situation where more than 20 requests are made
when draining the Q. to avoid that we are batching the scope together
in a single larger upload request and ensure that the sink queue is
flushed when full before adding new element and not losing any
symbols.