Add collection of local var for method probe
Collecting local var was disabled for method probe/unhanded exception
because all local vars are not accessible/defined.
But for Exception Debugging/Replay this is very useful to have.
The solution is to hoist local variable at the top level of the method
and initialize them to be able to capture them at any point of the
method.
we are handling also the case of duplicate local var in different
scope but with incompatible types