Fix couchbase flaky test (cleanup phase) (#7598)
Mute tracing on cleanup, instead of waiting for a cleanup span. This
test is flaky when it does not receive the cleanup span in time (under
heavy load).

Also, do not shallow exceptions on cleanup. If we ever get them, we can
address them, or add a more specific catch.