Add spark.stage details attribute at the end of the stage (#7608)
Adds the spark.stage details attribute at the end of the stage, rather than at the beginning.

The details attribute contains a large amount of data, including the full stack trace that initiated the stage. When using long-running spans, the span is flushed multiple times, which can significantly increase the ingestion volume.

By adding the details attribute at the end of the stage, we reduce the ingestion volume while still ensuring the information is available once the stage has completed.