Fix hoisting local vars for Kotlin code
javac begins the range of a local var in the LocalVariableTable just
after the store instruction at the init.
kotlinc is more free to instroduce nop/other instructions in-between.
need to scan backward from the start of the range to find the last
store of this slot to begin the rewrite of the slot needed to hoist
correctly local vars.