Avoid exception when capturing fields in jdk16+
Since jdk16, calling setAccessible(true) on fields that are on another
java module that is not open/exported will result with
InaccessibleObjectException thrown each time. To avoid this exception
we can call `trySetAccessible` that will return true if we can call
setAccessible without exception.
The method `trySetAccessible` is only available since jdk9 so we need
to use MethodHandle on it to be able to call it only when it is
available.
NB: setAccessible for jdk8 will always work anyway.
InaccessibleObjectException provides a useful message for the cause
of this error and we have added back a similar message by getting the
module where resides the field that is not accessible to explain to
the user why the field is not captured what it can do to resolve the
problem.
Fix also the report of evaluation errors for log template for
inaccessible fields inside an object (marked UNDEFINED)