Split SymDB payload when too large
SymDB payload cannot be larger than 50MB otherwise datadog agent
will generate an error. When serializing the upload request we are now
verifying that we are below this limit, otherwise we are splitting
the payload first by uploading by jar scope, but if a jar scope is
still large we are splitting by class scopes. If down to one class
the payload is still too large we are dropping the upload.