Add outer exceptions support for Exception Replay
In some cases, the innermost exception could contain only third-party
frames and no instrumentation is performed while the exception could
be re-throw up in the stack where the call originated. This could help
to root cause the inner exception.
Now in that case we are climbing up the chained of exceptions to find
user frames to instrument.
when finding the innermost exception, we are reporting the list of
all outer exceptions beginning by the innermost. The chained index is
also kept in the probe and injected into the snapshot to be able to
remap the frame index correctly and report it to the span as tag.