Avoid duplicate class symbol extraction (#7919)
Same classes can be loaded in different classloader, but symbol-wise
they are the same. In some situation (like Groovy script) we can a lot
 of class loaded for same symbols which pollute SymDB backend.
We need to track all classes loaded and extracted into ClassNameTrie
that is space efficient prefix tree. That way we prevent any
duplicates reaching SymDB backend