Add mechanism to remove exception probes
Add circuit breaker when evaluating exception probe to detect when
an exception probe is executed in a hot path only waiting for
exception to happen.
when circuit breaker is open after 1000 calls in 10 seconds window,
we check if there is no exception captured during this period too and
remove the probe from the configuration and un-instrument it.
This will prevent too much overhead for hot path