Capture values at entry for method probe (#8169)
method probe when evaluateAt is Exit and no condition should capture
entry values. When there is a condition, entry values are not captured
instrumentation check the presence of the condition to generate the
entry instrumentation or not. therefore, the dummy definition used
when duplicate probes need to merge the probe conditions if at least
one probe has one condition.
We are delegating the decision to capture through a new field in
CapturedContextInstrumentor