Remove flakiness from CustomLogManagerTest and CustomMBeanServerBuilder (#8241)
These tests historically set a dummy DD_API_KEY because that was required
by the profiler. However setting DD_API_KEY will now result in these tests
reaching out to the intake/backend because they won't be able to connect
to an agent and the fallback for certain features like telemetry is to try
and contact intake (these tests are not run alongside any test agent.)

When the call to intake/backend stalls this can lead to flaky timeouts.

We shouldn't be contacting the real intake during testing so this removes
the dummy api-keys from CustomLogManagerTest and CustomMBeanServerBuilder.

This is ok because profiling no longer needs DD_API_KEY set locally.