Introduce id-based line probe source matching
line probe in unit tests are created by searching probe id (UUID)
inside the targeted source file.
uuid are added as comment on the line we want to put a line probe.
Add helper method to search a uuid inside the targeted source file
and returns the line number used to create the line probe