Small improvement to agent feedback sampling
Previously, I updated this code to be case-insensitive.

In doing so, I introduced a call to String.toLowerCase which had a negative impact on response time and allocation.

By switching to TreeMap, I can use String::compareToIgnoreCase which avoids the allocation and has a better average complexity than toLowerCase.

This change provides a 1-1.5% improvement in a span creation throughput stress tests.