Avoid double snapshots for Exception Replay (#8273)
As Exception replay is supported by default for intermediate spans
we can end up in a situation where 2 spans are attaching the same
exception with the snapshots. in that case
DebuggerContext::handleException is called twice, once for each span
and we are also sending the same snapshots to the backend while only
one version is necessary and we tag it to the according spans.
We can just make sure we are sending the snapshots only once.