Rename debugger config to dynamic instrumentation
Originally the config parameter was debuggerXXX. when the product was
rolled out we use the term dynamic instrumentation. Now that we will
introduce the live debugger in itself, it's confusing to keep debugger
everywhere.
First step is to rename dynamic instrumentation config tokens to
reflect the fact that they are belonging to Dynamic Instrumentation
part of the product