change cookie vulns hash calculation (#8210)
What Does This Do
Change the evidence hash calculation for the location one

Motivation
Using evidence for the cookie vulnerabilities hash is not the most effective approach. In some applications, a different cookie name is used per request or session. This leads to a large number of duplicate vulnerabilities. Deduplicating by location leads to a predictably low and bounded number of vulnerabilities.