Fix CodeOrigin for @Trace annotation
add a param for captureCodeOrigin method to indicate if we need to
instrument for the CodeOriginProbe. the instrumentation for the
@Trace annotation insert a call to start and activate the span.
so we just need to call captureCodeOrigin after the activation so we
benefit from tracer instrumentation without generating a call to
CodeOriginProbe.
That way are in sync with the tracer instrumentation.
add smoke test