Fix Exception Replay with Lambda proxy classes (#8452)
If JVM is started with -XX:+ShowhiddenFrames lambda proxy classes
dynamically generated are shown in the stacktraces which may be used
in the fingerprinting for Exception Replay. The proxy class generated
contains an id that is different for each loading of the class.
Upon re-transformation this id is changing which led to a different
fingerprint for the same stacktrace which will trigger a new
instrumentation and a re-transformation. And again new fingerprint...
We are fixing this by filtering out lambda proxy classes if detected.