Reduce footprint of SourceFile tracking
classNamesBySourceFile map used for SourceFile tracking can have a lot
of entries (different SourceFiles) and for each entry we have a list
(If any inner classe or non-public top-level classes).
Compacting to a comma-separated list reduce the footprint of this map.
As the usage is very unfrequent we can pay the price to "deserialize"
the list to go through the classname that we need to retransform as a
dependency