Check known W3C propagation tags first when parsing (#8597)
This avoids triggering a misleading log message from TagKey, for example:

  DEBUG d.t.core.propagation.ptags.TagKey - Invalid header h: W3C s: dd=p:456789abcdef0123 b: 3 e: 4

This is not an invalid header, it's just that TagKey rejected the 'p:' last parent trace-id because it didn't start with the 't.' prefix for custom tags.
The code then falls back to checking known W3C propagation tags 's:', 'o:', and 'p:' - if we check for those first we can avoid the misleading message.