Document when we manage a span's lifecycle without scopes
There are a few instrumentations where the instrumented code does not align with scopes.
In such situations the span is activated, but the scope is thrown away. The span is then
closed in another section of code using 'closeActive()'.

This PR adds a 'activateSpanWithoutScope(span)' method to self-document this behaviour
and explicitly tie it to the use of 'closeActive()'.