Working on Entry test coverage
Since TagMap.Entry is designed to be multi-threaded -- working towards a multi-threaded test

Tests have now been restructured as a series of check function objects that are can be reordered to tests different orderings.

The test then repeated multiple times to test a variety of orderings.  This helps improve the test coverage as well.

Next step is to add multi-threading testing