Allow agent to be automatically injected when running aside Log4J patch agent (#8648)
# What Does This Do

* add an exception in double injection check for log4j patch agent

# Motivation

This fix is meant for DJM running in AWS EMR. 
AWS adds a javaagent on their spark processes to patch the log4j vulnerability. This wouldn't pass the current double injection check so the injector is currently configured to pass `DD_INJECT_FORCE` on EMR to force the check.

The issue is that this is configuration is currently passed to the injector in a legacy way we want to remove, so we wouldn't be able to skip checks. This exception is also, in general, a bit scary because it applies everywhere we will inject everything on EMR regardless of deny lists.

So this PR adds a more restricted exception in the AgentBootstrap, and only works if there 2 agents one of them being for log4j.