feat: Skip JAXB generated classes classloader
In a OSGI environment JAXB code generation creates a lot `com.ibm.xml.xlxp2.jaxb.codegen.AbstractGeneratedStubFactory$RootStubClassLoader` this classloader is used to define generated classes that must be actually defined in the correct OSGI bundle.

Given the tracer needs to instrument classloading and also perform osgi bundle lookup this interaction create a lot of load in the tracer when these generated classes are defined.

Given this classloader is used for the generated classes, it is safe to exclude skip it.